Added launch screen. Updated version number to 2.0. Fixed a number of warnings. I believe I fixed a new iOS9 runtime error that occurred at the end of computing/right when analysis controller was presented (I think it was due to device rotating)

+++ BWT_Matcher_InsertionDeletion.m
@@ -218,7 +218,7 @@
     //printf("DK: creating newa\n");
     
     //    Add space prior to the chars in "a" and prior to the chars in "b"
-    int alen = strlen(a);
+    int alen = (int)strlen(a);
     char *newa = calloc(alen+2, 1);
     memcpy(newa+1, a, alen);
     newa[0] = ' ';
